refactor: update article imports and enhance comment repository type handling
- Replaced `articlesPlugin` with `articlesController` in the app module for improved structure.
- Updated article schema imports across various files to maintain consistency.
- Refactored `CommentsRepository` to use a new `CreateCommentDto` type for better clarity and type safety.
- Cleaned up unused imports and comments for enhanced code readability.
