+++ src/comments/schema/comment-relations.schema.ts
@@ -1,4 +1,4 @@
-import { articles } from '@articles/schema';
+import { articles } from '@/articles/articles.schema';
 import { users } from '@users/users.model';
 import { relations } from 'drizzle-orm';
 import { comments } from './comments.schema';
