+++ src/users/users.model.ts
@@ -1,7 +1,4 @@
-import {
-  articles,
-  favoriteArticles,
-} from '@/articles/schema/favorite-articles.schema';
+import { articles, favoriteArticles } from '@/articles/articles.schema';
 import { relations, sql } from 'drizzle-orm';
 import {
   date,
